Staff Notes v0.0.1 StaffNotes.Accounts

Represents the business-logic layer of working with the various accounts in the Staff Notes application. There are three types of accounts in Staff Notes:

Organizations - Represents an organization that has members whose behavior and achievements need to be tracked by staff users

Teams - Represents a group of users with the same permission level in an organization

Users - Represents a staff member of an organization who will access or author notes

Business logic specific to a particular record is typically implemented within that module. For example, the logic around original teams is implemented in the StaffNotes.Accounts.Team module. Business logic that involves multiple records, such as when a user creates an organization (see the create_org/1 function) is implemented within this module.

Functions

add_user_to_org(user, org)

add_user_to_org(StaffNotes.Accounts.User.t(), StaffNotes.Accounts.Organization.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Adds the user to the organization.

add_user_to_team(user, team)

add_user_to_team(StaffNotes.Accounts.User.t(), StaffNotes.Accounts.Team.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Adds the user to the team.

change_org(org)

change_org(StaffNotes.Accounts.Organization.t()) :: Ecto.Changeset.t()

Creates an Ecto.Changeset for tracking organization changes.

change_team(team)

change_team(StaffNotes.Accounts.Team.t()) :: Ecto.Changeset.t()

Creates an Ecto.Changeset for tracking team changes.

Examples

iex> change_team(team)
%Ecto.Changeset{source: %Team{}}

change_user(user)

change_user(StaffNotes.Accounts.User.t()) :: Ecto.Changeset.t()

Creates an Ecto.Changeset for tracking user changes.

Examples

iex> change_user(user)
%Ecto.Changeset{source: %User{}}

create_org(attrs \\ %{}, user)

create_org(Map.t(), StaffNotes.Accounts.User.t()) ::
  {:ok, StaffNotes.Accounts.Organization.t()} |
  {:error, Ecto.Changeset.t()}

Creates an organization.

When an organization is created:

It is created by a user

The organization is created

An "owners" team is created for that organization (with "owner" permissions)

The creating user is added to the organization

The creating user is added to the "owners" team

All of this is executed in a database transaction so if any step fails, it all gets rolled back.

create_team(team_attrs \\ %{}, org)

create_team(Map.t(), StaffNotes.Accounts.Organization.t()) ::
  {:ok, StaffNotes.Accounts.Team.t()} |
  {:error, Ecto.Changeset.t()}

Creates a team within an organization.

See StaffNotes.Accounts.Team.create_team_changeset/2 for applicable business rules.

create_user(attrs \\ %{})

create_user(Map.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Creates a user.

Examples

iex> create_user(%{field: value})
{:ok, %User{}}

iex> create_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_org(org)

delete_org(StaffNotes.Accounts.Organization.t()) ::
  {:ok, StaffNotes.Accounts.Organization.t()} |
  {:error, Ecto.Changeset.t()}

Deletes the organization.

This also deletes all teams within the organization and disassociates all users from the organization and its teams.

delete_team(team)

delete_team(StaffNotes.Accounts.Team.t()) ::
  {:ok, StaffNotes.Accounts.Team.t()} |
  {:error, Ecto.Changeset.t()}

Deletes the team.

See StaffNotes.Accounts.Team.delete_team_changeset/1 for applicable business rules.

delete_user(user)

delete_user(StaffNotes.Accounts.User.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Deletes a user.

Examples

iex> delete_user(user)
{:ok, %User{}}

iex> delete_user(user)
{:error, %Ecto.Changeset{}}

get_org(name, options \\ [])

get_org(String.t(), Keyword.t()) ::
  StaffNotes.Accounts.Organization.t() |
  nil

Gets a single organization.

Works the same as get_org!/1 but returns nil instead of raising an error when the organization is not found.

get_org!(name, options \\ [])

get_org!(String.t(), Keyword.t()) ::
  StaffNotes.Accounts.Organization.t() |
  no_return()

Gets a single organization.

Raises Ecto.NoResultsError if the organization does not exist.

Options

:with — Specifies the records to preload in the organization record. It uses the same keyword list syntax as Ecto.Repo.preload/3

Examples

Gets just the organization record:

Accounts.get_org!("org-name")

Gets the organization record and preloads the users field.

Accounts.get_org!("org-name", with: [:users])

get_team!(name)

get_team!(String.t()) :: StaffNotes.Accounts.Team.t() | no_return()

Gets the team by id.

Raises Ecto.NoResultsError if the team does not exist.

get_team!(org_name, team_name)

get_user(name_or_id)

get_user(String.t() | integer()) :: StaffNotes.Accounts.User.t() | nil

Gets a single user.

Works the same as get_user!/1 but returns nil instead of raising an error when the user is not found.

get_user!(name_or_id)

get_user!(String.t() | integer()) ::
  StaffNotes.Accounts.User.t() |
  no_return()

Gets a single user.

Raises Ecto.NoResultsError if the user does not exist.

Examples

Get a user by user ID:

iex> get_user!(123)
%User{}

Get a user by their user name:

iex> get_user!("username")
%User{}

Attempt to get a non-existent user:

iex> get_user!(456)
** (Ecto.NoResultsError)

list_orgs()

list_orgs() :: [StaffNotes.Accounts.Organization.t()]

Returns the list of organizations.

Examples

iex> list_orgs()
[%Organization{}, ...]

list_teams(org)

list_teams(StaffNotes.Accounts.Organization.t()) :: [StaffNotes.Accounts.Team.t()]

Lists all teams within the organization.

list_users()

list_users() :: [StaffNotes.Accounts.User.t()]

Returns the list of all users.

Examples

iex> list_users()
[%User{}, ...]

list_users(org)

list_users(StaffNotes.Accounts.Organization.t()) :: [StaffNotes.Accounts.User.t()]

list_users(StaffNotes.Accounts.Team.t()) :: [StaffNotes.Accounts.User.t()]

List the users on a team.

original_team(organization_id)

original_team(String.t()) :: StaffNotes.Accounts.Team.t() | nil

Returns the original team for the organization, if it exists.

This shouldn't be susceptible to race conditions so long as the original team is always created along with the organization.

remove_user_from_org(user, org)

remove_user_from_org(StaffNotes.Accounts.User.t(), StaffNotes.Accounts.Organization.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Removes the user from the organization.

remove_user_from_team(user, team)

remove_user_from_team(StaffNotes.Accounts.User.t(), StaffNotes.Accounts.Team.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Removes the user from the team.

update_org(org, attrs)

update_org(StaffNotes.Accounts.Organization.t(), Map.t()) ::
  {:ok, StaffNotes.Accounts.Organization.t()} |
  {:error, Ecto.Changeset.t()}

Updates the organization.

update_team(team, attrs)

update_team(StaffNotes.Accounts.Team.t(), Map.t()) ::
  {:ok, StaffNotes.Accounts.Team.t()} |
  {:error, Ecto.Changeset.t()}

Updates the team.

See StaffNotes.Accounts.Team.update_team_changeset/2 for applicable business rules.

update_user(user, attrs)

update_user(StaffNotes.Accounts.User.t(), Map.t()) ::
  {:ok, StaffNotes.Accounts.User.t()} |
  {:error, Ecto.Changeset.t()}

Updates a user.

Examples

iex> update_user(user, %{field: new_value})
{:ok, %User{}}

iex> update_user(user, %{field: bad_value})
{:error, %Ecto.Changeset{}}