Staff Notes v0.0.1 StaffNotes.Accounts.Team

A team is a collection of users within an organization that have the same permission level.

Permission Level

There are three permission levels:

:owner

All permissions of :write

Can invite users to the organization

Can remove users from the organization

If there is only one owner, they can delete the organization

Can create or delete teams

Can rename teams

Can add or remove users from teams

Can view list of users within all teams

:write

All permissions of :read

Can create members

Can create identities

Can merge members

Can create notes

Can edit notes they have authored

:read

Can view notes

Can view members

Can view identities

Can view list of teams within the organization

Can view list of users within the organization

Can view list of users within teams to which the user belongs

There can be multiple teams with the same permission level so that an organization can track users how they wish.

Original

The original team is special:

There can be only one original team

It cannot stop being the original team

It is created with owner permission level

It cannot be deleted

It cannot have its permission level changed

If the original team contains one member, they are not allowed to leave the team

If the original team contains one member, they are not allowed to leave the organization

This is done so that there will always be at least one organization member that is capable of administrating the organization.

Types

t()

t() :: %StaffNotes.Accounts.Team{__meta__: term(), id: term(), inserted_at: term(), name: term(), organization: term(), organization_id: term(), original: term(), permission: term(), updated_at: term(), users: term()}

Functions

changeset(team, attrs)

changeset(t(), %{}) :: Ecto.Changeset.t()

Generates an Ecto.Changeset that is applicable to all database operations.

A more specialized changeset will implement further business rules and should be preferred.

create_team_changeset(team, attrs \\ %{})

create_team_changeset(t(), %{}) :: Ecto.Changeset.t()

Creates an Ecto.Changeset suitable for creating a new team.

Business rules

Prevents creating an original team if one already exists in the organization

delete_team_changeset(team)

delete_team_changeset(t()) :: Ecto.Changeset.t()

Creates an Ecto.Changeset suitable for deleting a team.

Business rules

Prevents deleting an original team

update_team_changeset(team, attrs \\ %{})

update_team_changeset(t(), %{}) :: Ecto.Changeset.t()

Creates an Ecto.Changeset suitable for updating a team.

Business rules

Prevents changing the value of the original field

Prevents changing of an original team's permission level