Staff Notes v0.0.1 StaffNotes.Ecto.Markdown

An Ecto.Type that represents a chunk of Markdown to be rendered.

Use this as the type of the database field in the schema:

defmodule StaffNotes.Notes.Note do
  use Ecto.Schema
  alias StaffNotes.Ecto.Markdown

  schema "notes" do
    field :title, :string
    field :text, Markdown
  end
end

This type requires special handling in forms because Phoenix's form builder functions call Phoenix.HTML.html_escape/1 on all field values, which returns the html field on this type. But what we want when we show an StaffNotes.Ecto.Markdown.t/0 value in a form is the text field.

See: Beyond Functions in Elixir: Refactoring for Maintainability

Types

markdown()

markdown() ::
  %StaffNotes.Ecto.Markdown{html: term(), text: term()} |
  String.t()

t()

t() :: %StaffNotes.Ecto.Markdown{html: nil | String.t(), text: String.t()}

Functions

to_html(markdown)

to_html(markdown()) :: String.t()

Renders the supplied Markdown as HTML.

Examples

Render Markdown from a string:

iex> StaffNotes.Ecto.Markdown.to_html("# Foo")
"<h1>Foo</h1>
"

Render Markdown from an unrendered Markdown struct:

iex> StaffNotes.Ecto.Markdown.to_html(%StaffNotes.Ecto.Markdown{text: "# Foo"})
"<h1>Foo</h1>
"

Passes already rendered Markdown through unchanged:

iex> StaffNotes.Ecto.Markdown.to_html(%StaffNotes.Ecto.Markdown{html: "<p>foo</p>"})
"<p>foo</p>"

Returns an empty string for anything that isn't a string or a Markdown struct:

iex> StaffNotes.Ecto.Markdown.to_html(5)
""

to_iodata(markdown)

Converts a block of markdown to iodata.