Staff Notes v0.0.1 StaffNotes.Ecto.Slug

An Ecto.Type that represents an identifier that can be used in a URL.

This type also makes it so that binary_id is distinguishable from a name because it keeps them both strongly typed.

Use

Use this as the type of the database field in the schema:

defmodule StaffNotes.Accounts.Organization do
  use Ecto.Schema
  alias StaffNotes.Ecto.Slug

  schema "organizations" do
    field :name, Slug
  end
end

Format

Follows the GitHub pattern for logins. They consist of:

Alphanumerics — /[a-zA-Z0-9]/

Hyphens — /-/

Must begin and end with an alphanumeric

Types

t()

t() :: %StaffNotes.Ecto.Slug{text: String.t()}

Functions

cast(slug)

Casts the given value into a slug.

dump(slug)

Dumps the given value into an Ecto native type.

load(binary)

Loads the given value into a slug.

to_iodata(slug)

Converts a slug to iodata.

to_string(slug)

Converts a slug to a string.

type()

Returns the underlying schema type for a slug.

valid?(binary)

valid?(t() | String.t()) :: boolean()

Determines whether the given value is valid to be used as a slug.