Staff Notes v0.0.1 StaffNotes.IEx.Helpers

Helper functions for use in the IEx console.

All of these functions are available directly at the interactive prompt with no namespace necessary.

Functions

dat(queryable)

Gets the first record that matches the given query.

Examples

dat Organization

demote_user(user)

demote_user(StaffNotes.Accounts.User.t()) :: StaffNotes.Accounts.User.t()

Demotes the given user to a normal non-admin account.

demote_user!(name)

demote_user!(String.t()) :: StaffNotes.Accounts.User.t() | no_return()

Demotes the named user to a normal non-admin account.

Raises an Ecto.NoResultsError if the named user cannot be found.

exit()

Cleanly shuts down the local node and exits the shell.

promote_user(user)

promote_user(StaffNotes.Accounts.User.t()) :: StaffNotes.Accounts.User.t()

Promotes the given user to a site admin.

promote_user!(name)

promote_user!(String.t() | StaffNotes.Accounts.User.t()) ::
  StaffNotes.Accounts.User.t() |
  no_return()

Promotes the named user to a site admin.

Raises an Ecto.NoResultsError if the named user cannot be found.