Staff Notes v0.0.1 StaffNotesWeb

The entrypoint for defining your web interface, such as controllers, views, channels and so on.

This can be used in your application as:

use StaffNotesWeb, :controller
use StaffNotesWeb, :view

The definitions below will be executed for every view, controller, etc, so keep them short and clean, focused on imports, uses and aliases.

Do NOT define functions inside the quoted expressions below. Instead, define any helper function in modules and import those modules here.

Types

params()

params() :: Map.t()

Parameters passed to a controller action.

Functions

__using__(which) (macro)

When used, dispatch to the appropriate controller/view/etc.

channel()

Adds alias, import and use directives for Phoenix.Channel modules.

controller()

Adds alias, import and use directives for Phoenix.Controller modules.

helper()

Adds alias, import and use directives for modules intended to create helper functions for Phoenix.View modules.

router()

view()

Adds alias, import and use directives for Phoenix.View modules.