Staff Notes v0.0.1 StaffNotesWeb.LayoutView

View functions for the site's main page layout.

Types

app_name()

app_name() :: atom()

The application name as an atom.

link_info()

link_info() :: {String.t(), String.t()}

The link information as a tuple of the author's name and URL to link to.

maybe_user()

maybe_user() :: StaffNotes.Accounts.User.t() | nil

Represents a type that may be a user or may represent nothing.

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

code_with_heart(link_info, options \\ [])

code_with_heart(app_name() | link_info(), Keyword.t()) :: Phoenix.HTML.safe()

Renders the GitHub-style <> with ♥ by [author link] footer item.

This function can read the author link information from the Application environment. You can set the author link information by adding the following to your config.exs:

config :app_name,
  author_name: "Author's name",
  author_url: "https://example.com"

Or you can supply the author link information as a {name, url} tuple as the first argument.

Options

Options are used to customize the rendering of the element.

:link_options -- A keyword list passed as attributes to the author link a tag

All other options are applied as attributes to the containing div element

Examples

Phoenix.HTML.safe_to_string(LayoutView.code_with_heart(:app_name))
#=> "<div><svg .../> with <svg .../> by <a href="https://example.com">Author's Name</a></div>"

github_link(url, options \\ [])

github_link(atom() | String.t(), Keyword.t()) :: Phoenix.HTML.safe()

Renders the link to the source repository.

Options

:text -- If :text is true, use GitHub as the link text; otherwise use the GitHub mark octicon (defaults to false)

All other options are passed to Phoenix.HTML.Link.link/2

login_button(conn, current_user)

login_button(Plug.Conn.t(), maybe_user()) :: Phoenix.HTML.safe()

Renders the appropriate login buttons depending on whether the user is signed in.

When current_user is nil, the login button is displayed. When current_user is defined, a logout link and link to the user's profile page is displayed.

render(template, assigns \\ %{})

Renders the given template locally.

render_flash(flash_info)

render_flash(Plug.Conn.t() | Map.t()) :: Phoenix.HTML.safe()

Renders the flash content.

Examples

render_flash(@conn)

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.