Staff Notes v0.0.1 StaffNotesWeb.NoteController

Handles requests for StaffNotes.Notes.Note records.

Functions

create(conn, map)

Receives parameters for creating a new note for an organization and stores it in the database.

delete(conn, params)

Receives a note to be deleted and deletes it from the database.

edit(conn, map)

Receives a note and renders a form for editing it.

index(conn, map)

Displays the list of notes for an organization.

new(conn, map)

Renders a form for creating a new note for an organization.

show(conn, map)

Displays a note.

update(conn, map)

Receives parameters for updating a note and saves it to the database.