Staff Notes v0.0.1 StaffNotesWeb.PrimerHelpers

Helper functions for generating elements that work with Primer.

Functions

avatar(user, options \\ [])

avatar(User.t(), keyword()) :: Phoenix.HTML.safe()

Displays the avatar for the StaffNotes.Accounts.User.

Options

Valid options are:

:size -- the value in pixels to use for both the width and height of the avatar image

input(form, field, options \\ [])

input(Phoenix.HTML.FormData.t(), atom(), keyword()) :: Phoenix.HTML.safe()

Displays the appropriate input control for the given field.

Options

:using -- override the built-in selection of input field based on data type. Can be any of the Phoenix.HTML.Form input function names or the special value :markdown which displays a specially-formatted textarea

See: Dynamic forms with Phoenix

link_button(text, options \\ [])

Generates a link button with the given text and options.

Options

required :to -- the URL to link to

render_list(collection, module, template_root, assigns \\ %{})

Renders a standard list of items.

Takes the same parameters and options as Phoenix.View.render_many/4 except that the template_root parameter is the root name of the template. This root name will have either _blankslate.html appended if collection is empty or have _item.html appended if the collection is non-empty.

tabnav(list)

Renders the tabnav element.

Examples

In Elixir code:

tabnav do
  [
    tabnav_item("Text", "https://example.com")
  ]
end

In Slime template:

= tabnav do
  = tabnav_item("Text", "https://example.com")

tabnav_item(text, path, options \\ [])

Generates a tabnav item to be rendered inside a tabnav element.

Examples

Rendering a tabnav item with an icon, counter, aligned right, and selected:

tabnav_item(
  "Settings",
  "https://example.com",
  counter: 5,
  icon: :gear,
  right: true,
  selected: true
)