Staff Notes v0.0.1 StaffNotesWeb.Router.Helpers

Module with named helpers generated from StaffNotesWeb.Router.

Functions

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

organization_note_path(conn_or_endpoint, action, organization_name)

organization_note_path(conn_or_endpoint, action, organization_name, params)

organization_note_path(conn_or_endpoint, action, organization_name, id, params)

organization_note_url(conn_or_endpoint, action, organization_name)

organization_note_url(conn_or_endpoint, action, organization_name, params)

organization_note_url(conn_or_endpoint, action, organization_name, id, params)

organization_path(conn_or_endpoint, action)

organization_path(conn_or_endpoint, action, name)

organization_path(conn_or_endpoint, action, name, params)

organization_staff_path(conn_or_endpoint, action, organization_name)

organization_staff_path(conn_or_endpoint, action, organization_name, params)

organization_staff_url(conn_or_endpoint, action, organization_name)

organization_staff_url(conn_or_endpoint, action, organization_name, params)

organization_team_path(conn_or_endpoint, action, organization_name)

organization_team_path(conn_or_endpoint, action, organization_name, params)

organization_team_path(conn_or_endpoint, action, organization_name, name, params)

organization_team_url(conn_or_endpoint, action, organization_name)

organization_team_url(conn_or_endpoint, action, organization_name, params)

organization_team_url(conn_or_endpoint, action, organization_name, name, params)

organization_url(conn_or_endpoint, action)

organization_url(conn_or_endpoint, action, name)

organization_url(conn_or_endpoint, action, name, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action, name)

user_path(conn_or_endpoint, action, name, params)

user_url(conn_or_endpoint, action, name)

user_url(conn_or_endpoint, action, name, params)