Staff Notes v0.0.1 StaffNotesWeb.TeamController

Handles requests for StaffNotes.Accounts.Team records.

All actions are for teams within an organization. The organization is identified by the organization_name parameter.

Functions

create(conn, params)

Receives parameters for creating a new team for an organization and stores it in the database.

delete(conn, params)

Receives a team to be deleted from an organization by name and deletes it from the database.

edit(conn, params)

Receives a team within an organization by name and renders a form for editing it.

index(conn, map)

Displays the list of teams for an organization.

new(conn, params)

Renders a form for creating a new team for an organization.

show(conn, map)

Displays a team for an organization.

update(conn, params)

Receives parameters for updating a team within an organization and saves it to the database.