Staff Notes v0.0.1 StaffNotesWeb.TimeHelpers

View helpers for dealing with timestamps.

Functions

timeago(time)

timeago(NaiveDateTime.t()) :: Phoenix.HTML.safe()

Uses the timeago.js module to format the given timestamp as "[some amount of time] ago" such as "three hours ago".

HTML emitted

It emits an HTML span tag of the format:

<span class="timeago" datetime="2018-01-01T12:00:00Z" title="2018-01-01T12:00:00Z">
  2018-01-01T12:00:00Z
</span>

The timeago class is used by the library to find content to reformat based on the datetime property. The title property is included so that people can hover over the time expression to see the exact date and time the event occurred. The content of the span tag is there in case the viewer has JavaScript disabled in their browser. The content of the tag is replaced otherwise by the timeago library.

Time zones

This function assumes that all timestamps are in UTC.

Examples

Slime template:

= timeago(~N[2018-01-01 12:00:00])