Staff Notes v0.0.1 StaffNotesWeb.UserView

View functions for the user pages.

Functions

__phoenix_recompile__?()

Returns true whenever the list of templates changes in the filesystem.

__resource__()

The resource name, as an atom, for this view

__templates__()

Returns the template root alongside all templates.

create_org_button(conn, options \\ [])

Renders a "Create organization" button.

render(template, assigns \\ %{})

Renders the given template locally.

render_org_info_list(conn, user)

Renders the list of organizations the given user belongs to.

show_org_link(conn, org)

Creates a link to the given organization's show page.

staff_badge(user, options)

staff_badge(StaffNotes.Accounts.User.t(), Keyword.t()) :: Phoenix.HTML.safe()

Displays a staff badge if the given user is a site admin.

template_not_found(template, assigns)

template_not_found(Phoenix.Template.name(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.